pyproject.toml
[build-system]
requires = ["setuptools>=77"]
build-backend = "setuptools.build_meta"

[project]
name = "compass-a2a"
version = "0.1.0"
description = "A2A adapter service for Compass"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "liujuanjuan1984" }]
dependencies = [
    "a2a-sdk[http-server]==0.3.25",
    "fastapi>=0.110,<1.0",
    "pydantic-settings>=2.2,<3.0",
    "sse-starlette>=2.1,<4.0",
    "uvicorn>=0.29,<1.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27,<1.0",
    "pre-commit>=4.0,<5.0",
    "pytest>=8.0,<10.0",
    "ruff>=0.5,<0.16",
]

[project.scripts]
compass-a2a = "compass_a2a.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
package = true
